Recommitting this patch on the corrected 2.6 branch.
This commits the relevant changes from Issue 143, including the refactored CM and the newer testcases to support better CM testing.  A few tests fail in the newer AbstractCmTest against the default BTree implementation, due to lack of some index operations.  These need a new feature request.

Not included here are is the BDB implementation which was submitted with the original patch, nor the fixes to BASE64Test, which work great, but need a separate issue.